fix(set): Fix set directives not streaming correctly on Node 16
