+++ packages/astro/test/streaming.test.js
@@ -1,7 +1,7 @@
-import { isWindows, loadFixture } from './test-utils.js';
 import { expect } from 'chai';
-import testAdapter from './test-adapter.js';
 import * as cheerio from 'cheerio';
+import testAdapter from './test-adapter.js';
+import { isWindows, loadFixture, streamAsyncIterator } from './test-utils.js';
 
 describe('Streaming', () => {
 	if (isWindows) return;
@@ -32,7 +32,7 @@ describe('Streaming', () => {
 		it('Body is chunked', async () => {
 			let res = await fixture.fetch('/');
 			let chunks = [];
-			for await (const bytes of res.body) {
+			for await (const bytes of streamAsyncIterator(res.body)) {
 				let chunk = bytes.toString('utf-8');
 				chunks.push(chunk);
 			}
@@ -61,7 +61,7 @@ describe('Streaming', () => {
 			const response = await app.render(request);
 			let chunks = [];
 			let decoder = new TextDecoder();
-			for await (const bytes of response.body) {
+			for await (const bytes of streamAsyncIterator(response.body)) {
 				let chunk = decoder.decode(bytes);
 				chunks.push(chunk);
 			}
@@ -102,7 +102,7 @@ describe('Streaming disabled', () => {
 		it('Body is chunked', async () => {
 			let res = await fixture.fetch('/');
 			let chunks = [];
-			for await (const bytes of res.body) {
+			for await (const bytes of streamAsyncIterator(res.body)) {
 				let chunk = bytes.toString('utf-8');
 				chunks.push(chunk);
 			}
