+++ packages/astro/test/test-utils.js
@@ -303,3 +303,17 @@ export const isWindows = os.platform() === 'win32';
 export function fixLineEndings(str) {
 	return str.replace(/\r\n/g, '\n');
 }
+
+export async function* streamAsyncIterator(stream) {
+	const reader = stream.getReader();
+
+	try {
+		while (true) {
+			const { done, value } = await reader.read();
+			if (done) return;
+			yield value;
+		}
+	} finally {
+		reader.releaseLock();
+	}
+}
